Python library for managing the lbrynet client from LBRY
The Python library is in `lbrytools`. The entire directory (package)
must be installed in a path searched by Python (`site-packages`),
then it can be imported.
```
import lbrytools
```

The `zeedit` directory contains the `zeedit.py` script that uses
the `lbrytools` methods to download claims from the LBRY network.
It requires the configuration file `zeedit_config.py` to set
the variables for the `lbrytool` functions.
```
python zeedit.py
```

To have a self-contained solution, the entire `lbrytools`
directory can be placed next to the `zeedit.py` script and
`zeedit_config.py`.

The `lbrynet` daemon must be running before using these tools.
This program can be started by itself.
```
lbrynet start
```

Otherwise it runs in the background when the LBRY Desktop application
is opened.