Clean: calculate used space with shutil.disk_usage
Instead of using `subprocess` with `du -s` we use
`shutil.disk_usage` to be platform independent.

The older `du -s` provides a value in Kibibytes (KiB)
so it needs to be divided by (1024^2) to get Gibibytes (GiB).
On the other hand `shutil.disk_usage` provides a value in bytes,
so it needs to be divided by (1024^3) to get GiB.