Download: randomize the processing of channels
Normally with `ch_download_latest_multi` the list of `channels`
is processed sequentially, as the channels are ordered.

If the list is very long the LBRY daemon may stop processing
the claims, so it may happen that only the first channels
are processed but not the last ones.

By processing the channels in random order, we increase
the probability of processing all channels by running this function
`ch_download_latest_multi` several times.