Zeedit: add option to shuffle the list of channels
In the configuration file set `suffle` to 1 (True)
to shuffle the list of channels so that they aren't always
downloaded in sequence. Occasionally the LBRY daemon stops when many
downloads are active, so this is useful to make sure all channels are
considered, and not only the first ones in the list.

This option is passed to `ch_download_latest_multi` in `zeedit.py`.