Download: rename `redownload_claims` to `download_claims`
The function `redownload_claims` makes more sense named
`download_claims`, as its primary task is to download claims
from a file that has many claims. If these claims already
exist, then they will be redownloaded, or the missing blobs
will be completed.

The `lbrytools.md` documentation file is augmented with a
new section "Download from file" to more correctly explain
how to use this function. The information was moved
from the "Re-download" section.

This renaming would break scripts that previously used
the older name but does not break `zeedit.py`,
as it is not used there.

The README.md is updated accordingly.