print: skip printing empty item in `print_multi_list`
This function is used to process the output of
`download.ch_download_latest_multi`.
However, if the download fails for certain claims,
their corresponding elements in the output list may be `False`.
So we detect this condition and skip the item in the loop.

Also fix some small indentation issues to keep the lines
aligned.