search: warning about searching for channel with `'channel_name'`
Currently searching by channel using
```
lbrynet file list --channel_name=@Channel
```
does not work very well because only a small portion of channels
are searchable.

A solution to this problem would be to get the list of all items,
and then perform a reverse lookup to find the channel
of each item, although this search is quite slow.

As it is right now, `sort_items` (and thus `lbrynet file_list`)
accepts any channel name, whether it exists or not.
Now, if there are no items found, a message indicates that it may be
due to the channel not existing or being mispelled.