search: new function `resolve_channel` to find a channel
Using the `lbrynet resolve <url>` command can be used to resolve
claims but also channels if the `<url>` starts with the `@` symbol.

Resolving a channel is necessary so that other functions
are able to find the channel, for example
```
lbrynet file list --channel_name=@Channel
```

This function is used by `sort_items(channel=channel)`,
and will be used by `ch_search_latest` in a future commit.