download: resolve the channel name before `lbrynet get`
A bug (lbryio/lbry-sdk #3316) prevents the
`lbrynet file list --channel_name=@Channel`
command from finding the channel, therefore the channel must be
resolved with `lbrynet resolve` before it becomes known by other
functions.

This is done by running `resolve_channel` inside `lbrynet_get`
which is then used by `download.download_single`,
and all the other functions that use it.
This will resolve the channel at the moment a claim from it
is downloaded.

The internal download function `lbrynet_get`
returns a dictionary with various types of information,
including the `'channel_name'`.
If the channel hasn't been resolved, the value of `'channel_name'`
will be `None`, and won't be displayed correctly by functions
that use this dictionary, like `print.print_multi_list`
which reads the output from `download.ch_download_latest_multi`
in `zeedit.py`.

Both the short `@Name` and the canonical `@Name#7` are resolved.
The second form is necessary to get the exact channel, in case
it has the same base name as another channel.

The download directory will now use the canonical name
but with ':' instead of '#', like `@Name:7`.
The reason is that some shells interprete the `#` symbol
in a special way, so avoiding this character removes
the need to escape it in some way, for example, `\#`.