search: small changes to the error detection and error messages
In `search_item_uri`, `search_item_name`, and `resolve_channel`
instead of a simple message to 'check spelling',
a longer message is shown and whether the input was `uri`, `claim id`,
`name`, or `channel`.

In `sort_items`, we count the items and return earlier, if the number
of items is zero, as this may indicate no downloaded items,
or a non-existent channel.

We also print a message if during sorting we are using the `'timestamp'`
instead of the `'release_time'`. Some older claims, from before 2020,
did not have `'release_time'`.