search: print the information in the `search_item` method
The internal functions `search_item_uri` and `search_item_name`
are wrapped by `search_item` so this one is the one
printing the final error message.

Place the arguments each in its own line so it is clear what
is being passed to the function.

download: do not print error message after `search_item`
as this message is printed inside `search_item` already.