blobs: new function `blob_get`
This function is used inside `blobs_action` to either get a blob
from the LBRY network, or announce a blob.

At the moment it cannot be used with missing blobs;
the command hangs and never timeouts.
It can only be used with a blob that is already downloaded.

This bug is reported in lbryio/lbry-sdk, issue #2070.

Therefore, at this moment this function is not very useful.