blobs: improve the return value of `count_blobs`
If the claim is not found by `search.search_item` the returned
dictionary has the key `'error_not_found'`, as well
as the input values `'canonical_url'`, `'claim_id'`, and `'name'`.

This usually means that there was a spelling mistake with the URI
or the claim ID, or that the claim was removed by its author
from the network. Inspecting the blockchain will confirm if the LBC
spent in that claim was unspent.

If on the other hand the item exists but the corresponding `'sd_hash'`
is not in the `blobfiles` directory, the returned dictionary
will have the key `'error_no_sd_hash'`, as well
as the information about the item `'canonical_url'`, `'claim_id'`,
`'name'`, `'channel'`, and `'sd_hash'`.

This usually means that the file was downloaded at least once
but then the manifest blob `'sd_hash'` was somehow deleted.
The file just needs to be downloaded again to get this blob,
as this blob contains information on all other data blobs.

The `count_blobs` method is used inside `count_blobs_all`,
and the new outputs are handled in five cases: complete blobs,
incomplete blobs, no `sd_hash` blob, invalid claim,
and other types of errors.

The `count_blobs` method is also used inside `redownload_blobs`,
and the new outputs are handled. If the claim is not found
the method simply returns. If the claim does not have
the `sd_hash` blob it just prints a message but continues;
the same when other blobs are missing. As before, the function
returns if all blobs are already present, as nothing needs
to be downloaded.

The `count_blobs` method is also used inside `blobs_move`,
and the new outputs are handled. If the claim is not found
or if the `sd_hash` blob is missing the method will return
as there are no blobs to copy or move.