search, print: add newline and remove line in certain functions
Add a newline inside `print.print_items` after `search.find_channel`
if no channel is found.
If the newline is not provided the error messages will be
too cluttered.

Remove the `print` line after `search.search_item`
in `search.find_channel`.
The message is not necessary because `search_item`
already provides a message if the claim is not found.