clean: add single quotes around the name of a claim
This is to escape special characters, like parentheses
or hash symbols, if the `lbrynet` command
is copied and used directly in the terminal.
```
lbrynet file delete --claim_name='some-claim-name'
```

It will not work in every case, as for certain names
other characters may need to be escaped in different ways,
in particular if the name itself has single quotes.