print: in `print_info_pre_get` handle 'duration' of the claim
A 'video' or 'audio' claim may have a 'duration' value in seconds
in their dictionary.

In the past we always assumed videos had a 'duration'
but with further testing we notice this is not the case;
some claims do not have it.

We didn't test for 'duration' with audio files, but now we do
and display it accordingly.