search: new `invalid_claims` to get claims that are invalid
The `sort_items` method checks the output of `lbrynet file list`
which returns the local metadata of previously downloaded files.

Each item is then processed with `search_item` which internally
looks up the claim with `lbrynet resolve` or `lbrynet claim search`.
These commands check the online database (blockchain)
to get more information on the claim.

However, if the claim is not found online, it probably means that
the author decided to remove the claim after we downloaded it
originally. This can be verified with the blockchain explorer,
by following the claim ID for an 'unspent' transaction.

Once we know that a claim has become invalid online,
we may decide to do something about it, such as save its media file
or delete it completely together with its blobs,
in order to save hard disk space.