print: new `invalid` parameter for `print_summary`
It writes a file with all claims that are considered 'invalid',
that is, from the output from `search.sort_invalid`.

These are claims that were downloaded at some point, but which now
cannot be resolved anymore from the online database (blockchain).
This probably means that the author decided to remove the claims
at some point after they were downloaded originally.
This can be verified with the blockchain explorer, by following
the claim ID for an 'unspent' transaction.

The `invalid` parameter forces the use of `ch_online=False`
so that in combination with `ch=True` or `channel=name`,
it can resolve the names of the channels from the offline database
instead of trying to look up the name of the channels online
(which won't be found anyway).