print: new `ch_online` argument to resolve the channel's name
In `print_items` with `ch_online=True` during the loop of the items
it will use `search.find_channel`, which uses `search.search_item`,
to query the online database for the `'signing_channel'` of the
corresponding claim.

If `ch_online=False` it will simply return the `'channel_name'`
stored in the dictionary being processed.

This is necessary for those invalid items that have been
removed from the network, as their information can no longer
be queried online.

Also add the argument to `print_summary` which uses
`print_items`.