search: `print_error` argument to avoid printing errors
The method `search_item` normally prints messages when the
canonical url, claim ID or claim name is not found in the online
database (blockchain).

The `print_error=False` argument allows us to suppress
these messages. This is useful when this method is used
inside other functions when we want to limit the terminal output.