print: rename `date` to `fdate` as it's the file date
The `print_items` (and `print_summary`) methods always print
the formatted `'release_time'` of the processed claims.

The parameter `date` now is renamed to `fdate` to clarify
that this is the date added to the file name of the summary,
that is, the date is added to the `file` argument.