lbrytools.md: update the documentation on `print_summary`
Update the documentation to reflect the rename of `date` to `fdate`.

Indicate how to use the new `ch_online` parameter.
If `ch_online=True` it will be slow because it has to resolve
the claim online first.

If `ch_online=False` it will be fast because it uses the local
database. However, this may not work in every situation,
particularly if the channel hasn't been resolved previously.