clean: new `offline` parameter for `delete_single`
The `delete_single` method internally uses the `search_item`
method to obtain more information about the item being deleted.
By default `search_item` searches the online database for the claim
but this will fail if the claim is 'invalid', that is,
if it was already deleted online by its author.
In this case `delete_single` will return early and the local claim
(media file, blobs or both) won't be deleted.

By using `offline=True` internally `search_item` will also use
this parameter to query the local database, and will allow
the function to continue until the local claim is deleted.
This is required to delete these 'invalid' claims,
and make sure their blobs are completely removed.