clean: add `invalid` parameter to `remove_claims`
The `invalid` parameter is passed to the internal `delete_single`
function, in order to be able to delete a list of claims
that are 'invalid'.

The list of claims is a CSV document that can be generated
with the `print_summary` method also with the `invalid` parameter.
```
p = lbryt.print_summary(file="summary_invalid.txt", invalid=True)

q = lbryt.remove_claims(file="summary_invalid.txt", invalid=True)
```