clean: rename the parameter `offline` to `invalid`
For the internal command `search.search_item` the parameter
is called `offline` because it controls whether it searches
a claim ID or claim name offline or online.

However, for the `delete_single` function we want to be more
clear that the `invalid` parameter is used to handle
invalid claims, which are no longer in the online database.