download: new method `lbrynet_save` to recreate the downloaded file
This method can be used on claims for which we have all blobs
in order to recreate the media file (mp4, mp3, mkv, etc.).

It is mostly intended to be used with 'invalid' claims, that is,
those which have been removed from the online database (blockchain),
and thus cannot be redownloaded.

Once the media file is recreated, the binary blobs can be eliminated
as the claim can no longer be shared and seeded in the network.