print: augment `print_info_pre_get` with `offline` parameter
The `offline` parameter defaults to `False`, in which case
it assumes the item was resolved online through `lbrynet resolve`
or `lbrynet claim search`.

If it is `True` it assumes the item was resolved offline
through `lbrynet file list`.
This is required for 'invalid' claims, that is, those that
have been removed from the online database, so they only exist
locally.