blobs: new method `analyze_blobs` to provide information on the blobs
This function wraps around `count_blobs_all` to count all blobs
from all claims downloaded previously, or by a particular
channel, or range of indices.

It counts the files actually present in the `blobfiles` directory,
and compares them to the information provided by `count_blobs_all`.

It defines five categories for the claims:
claims with all their blobs,
claims with missing blobs (continue the downloads),
claims with `'sd_hash'` missing (redownload the claims),
claims not found in the network (invalid claims),
and claims with other errors.

It also prints a summary of the blobs, indicating whether
the blobs that we expect are present or not; this is given
by the sum of the data blobs and `'sd_hash'` blobs.

It checks whether there is an excessive amount of files
in the `blobfiles` directory, which probably means there are
orphaned blobs from claims that have become invalid.