blobs: small additions to `count_blobs_all`
Print a different message if the `channel` is provided and actual
items are found for it.

Also count the number of blobs that should exist for the claims
being processed. These blobs correspond to the claims
that are either complete or incomplete. We don't consider the claims
for which the `'sd_hash'` is missing, or invalid claims
which can be found online, since we don't know how many blobs
should exist.