blobs, search: remove `check_lbry` from functions
This method, `funcs.check_lbry`, is meant to check that the
`lbrynet` daemon is running, and if it is not running,
it starts it with `lbrynet start`.

However, at the moment it does not work properly.
If it is not started it will start it, but then it will block
the terminal, and it won't be possible to continue execution
of the code after the check.

So we remove it from the methods that still use it
until we can properly launch `lbrynet start`
without blocking the rest of the code.