print: new `print_channels` function to show all unique channels
It goes through the list of all downloaded claims and finds
the channel name for each of them, either in the online database
(default) or in the offline database.

Invalid channels are not reported if the search is online.
If invalid channels are searched offline, but their names
have not been resolved, they may not be reported either.