blobs: new `print_channel_analysis` method
It calls `print.print_channels` to get all channels
in the system.

Then it iterates on all channels, and calls `analyze_channel`
on each of them to get the number of claims, the number of blobs,
and the space that all those blobs take on the system.
This is done for both completed claims and for incomplete claims.

The information is summarized in tabular form and then printed
to the terminal or to a file.