clean: modify slightly the code for `pr_bar`
Instead of printing a single `msg` list of three strings,
we create three strings `limit_mark`, `bar`, and `usage_mark`,
and then put them together at the end in a list.

We also calculate the bar in a more compact way depending if
the usage is equal or larger than 99%.
In the past, the case for exactly 100% was not properly handled
as it would print an additional `=`.

Now the printed bar is always the same length.
The limiting characters at the ends are fixed, and only
the internal size is variable.