print: print the time used to run `print_channels`
It helps us see how long it takes to parse
the entire list of items, find the channel for each claim,
and then print the list.

New parameters `start` and `end` control how many claims
are inspected from the list of claims.

The `print_msg` parameter controls if the time
is actually printed.

Also add an empty line after not finding a channel
because otherwise the error messages don't have proper spacing.