download: change `:` to `_` as separator in the subdirectory
Windows doesn't seem to like `#` or `:` in the subdirectory
name when creating a subdirectory per channel, with `own_dir`.
This may also cause errors with shared drives (NAS, Samba, etc.).

In Linux it doesn't seem to be a problem, but we still change it
to improve the situation for everybody.