download: move various functions to `download_multi.py`
The functions that download a single file remain in `download.py`:
- `lbrynet_get`
- `download_single`
- `lbrynet_save`
- `download_invalid`

The functions that download multiple files are placed
in `download_multi.py`:
- `ch_download_latest`
- `ch_download_latest_multi`
- `redownload_latest`
- `download_claims`

The word "item" is replaced by "claim" in a few cases when it is
printed to the terminal.

This new module imports the old `download.py`,
and is also imported in `__init__.py`.