search: in `parse_claim_file` skip empty and comment lines
During the parsing we strip the current line of whitespace
and endline character `\n`. If the remaining line is empty,
we skip the line.
If the line starts with a hash `#`, like a comment,
we also skip it.

This prevents some errors, for example, when empty lines
exist at the beginning or end of the file.