zeedit: rename example configuration file `zeedit_config_example.py`
By default the script `zeedit.py` will try to load
a configuration file named `zeedit_config.py`.

Since we provide an example configuration file with this name,
it will be loaded. However, we may not want this because
it will create folders in our home directory, and potentially
download a few files.

Therefore, we will rename the configuration file to
`zeedit_config_example.py`, so that it won't be loaded automatically.

If anybody wants to test the script, the user can pass it
as first argument:
```
zeedit.py zeedit_config_example.py
```

Or the user may wish to rename it to the default name:
```
mv zeedit_config_example.py zeedit_config.py
```

Which then will be found automatically by the script when
no input is used.
```
zeedit.py
```