zeedit: new `z_defaults` module with default values for zeedit
The `z_defaults` method accepts as input a loaded module,
which in this case is the configuration module `zeedit_config.py`
loaded through the `zeedit.py` script.

The `z_defaults` method is run to verify that the configuration
has all necessary variables. If they are not defined, then it
will set defaults.

The only variable not set is `channels`, as this is necessary
to know which channels will be dowloaded.