print: include the blobs completed and total blobs in summary
At the end of the downloading step (1) in the `zeedit.py` script,
we print a short summary of the claims downloaded.

In this summary it is good to also print the number of blobs
of each claim to quickly see if the claim is complete
or if we need to redownload it, that is, run `zeedit.py` again.