print: print the list of claims only at the end
Previously, the loop would print each line during each cycle.
However, this doesn't always produce a clean output
because some items may be missing and need to be skipped,
which outputs a message on screen.

Instead, the string to be printed is collected inside
a list of strings, which is finally printed to the screen
or to the given file after the main loop has ended.