claims: new module `claims.py` with a function to check our own claims
The `claims_bids` method compares each of our claims against other
claims in the network that share the same claim `'name'`.
It specifically compares the bidding amount of LBC to see if we
have the "controlling" `'name'`.

We can filter by channels, ignore reposts, and ignore the claims
for which we already have the highest bid.

We can display the summary of claims on the terminal
or print it to a file.