search, download: add `sep` parameter to control the separator to parse claims
It is the separator character between the data fields
in the file that is parsed.
Since the claim name can have commas, a semicolon `;` is used
by default.

The `sep` parameter is added to the `search.parse_claim_file`
method, and also to the `download_multi.download_claims` method
which effectively downloads the items in the file.