zeedit: add `sm_sep` parameter to control the separator
Specify the `sm_sep` variable in the configuration file
for the `zeedit.py` script.
This is passed to the `sep` argument of the `printf.print_summary`
to indicate the separator that is used to separate the data fields
in the summary.

The default value is also specified in `z_defaults`.
Since claim names can have commas, a semicolon `;` is used
by default.