download: new `save_file` option for downloading a claim
In a recent commit to the lbry-sdk code, `lbrynet get`
sets the option `--save_file=False` by default in the settings.
This means only the first blob (`sd_hash`) in the stream will be saved
to the computer.

In order to trully download the rest of the blobs and create
the final media file, the `--save_file=True` option must be provided.

This option is provided and given to `download.lbrynet_get`
and all the methods that depend on it, like `download.download_single`.