print: the `'value_type'` is only present in items resolved online
The `print.print_info_pre_get` method is used on the item
found from the `search.search_item` method.

When the item is searched online it will have a `'value_type'` key,
as the object may be a stream, a repost, a collection, a channel,
or another type.

When the item is searched offline, among the downloaded files,
it does not have `'value_type'` but we assume it is a `'stream'`
because only streams can be downloaded.