zeedit: add `save_file` parameter to control the download of streams
Specify the `save_file` variable in the configuration file
for the `zeedit.py` script.
This is passed to the `save_file` argument of the `download.ch_download_latest_multi`
method to indicate whether to effectively download all blobs
of all streams, and create the media file (mp4, mp3, mkv, etc.)
in the download directory `ddir`.

The default value, `save_file=True`, is also specified
in `lbrytools.z_defaults`.