funcs: new function `server_exists` to verify that the daemon is running
It returns `True` if the server is running and `False` otherwise.

This function will be used at the beginning of most functions
that should connect to the `lbrynet` daemon, to verify that
it is running. If it is not running, it will print an error
message and return.

It is also used in `zeedit.py` so it doesn't continue if `lbrynet`
isn't running from the beginning.