blobs_ratio: new method `print_blobs_ratio` to show upload/download ratio
The blob seeding ratio is estimated from the number of times that the words
`"sent"` and `"downloaded"` are found in the log files of `lbrynet`,
normally under `~/.local/share/lbry/lbrynet`.

The most recent log is called `lbrynet.log`, then they go back
in time `lbrynet.log.1`, `lbrynet.log.2`, etc., until the ninth.

Each time the keywords are found, it is assumed one blob was uploaded
or downloaded from the peer to peer network. However, currently
there is no way of telling which blob it was, so it doesn't say
if it's uploading or downloading a variety of blobs, or only a limited
set.

Together with the number of blobs recorded, we also note their time.
In this way we can tell how many were uploaded or downloaded
in a particular timeframe, say, days, weeks, or months.

This information is shown in a global summary, and it is also broken up
per log file.

Finally, if Numpy and Matplotlib are available, the global time lists
are passed to a plotting function in order to plot histograms
of blobs processed (uploaded or downloaded) per hour
versus the day they were processed, counting back from today.