claims: rename option `skip_max` to `skip_controlling`, it makes more sense
With `skip_controlling` we will print all our claims except for
the ones we already control, that is, that have the highest bid.

When using `compact=True`, the values of the fields are adjusted
so that they line up better.

We also add new options to make the printing of certain fields
optional:
- show_claim_id: print the claim ID (40 characters)
- show_repost_status: print whether the claim is a repost or not
- show_competing: print the number of competing claims that share the same name
- show_reposts: print the number of reposts that share the same name

Also add the separator option `sep=;`.