support: new `abandon_support` to remove or change the support
This is essentially a wrapper over the `support_abandon` method.

If "claim" has 100, and we run `abandon_support("claim")`,
the new support will be 0.

However, we can input any number as `keep` to change our support,
either decrease or increase.

If "claim" has 100, and we run `abandon_support("claim", keep=50)`,
the new support will be 50.

If "claim" has 100, and we run `abandon_support("claim", keep=400)`,
the new support will be 400.

This method only works if there is an existing support on the claim.