support: new `target_support` to push a specified claim to have a certain support
If "claim" has 100, and we run `target_support("claim", target=500)`,
we will support that claim with an addition 400 so that the total
support is 500.

This can also be used to subtract an amount.

If "claim" has 500, and we run `target_support("claim", amount=400)`,
we will remove our support by 100 so that the remaining support is 400.

Each claim has a base support that is provided by the author
and by other users. Our specified `target` cannot be smaller than this value.
If we specify a lower value, all of our existing support, if any,
will be removed.
```
target = base + our_support
target = (original_publish + others) + our_support
```