support: do not display the supports for items that aren't resolved
The `support_list` shows the claims that are being supported.

These claims are then resolved to be able to obtain the support information.
However, if the hub servers are not responding correctly, we may not
obtain a resolved result.

The `resolved` list will only have those items that were actually
resolved.