download: modify the `save_file` option to exclusively download blobs
This option is given to `download.download_single`, and is passed
to the internal `lbrynet_get` method.

If `save_file=False` first the method calls `lbrynet get --save_file=False`
to download the first blob, which is the manifest blob (sd_hash).
Once this blob is downloaded, we use a GET request with
the `'streaming_url'` of the claim to get the remaining blobs
but no media file.
```
requests.get(claim["streaming_url"])
```

If `save_file=True` we call `lbrynet get --save_file=True`
to get all blobs, and create the media file with a single call.