claims: new `print_trending_claims` with various options
It will search claims, and we can specify `claim_type`
(stream, channel, repost, collection, livestream),
and `stream_types` (video, audio, document, image, binary,
and model).

The main function is `w_claim_search`, which is a wrapper
around `claim_search`.
When given the option `what=trending` it will search
for trending claims.

When given the option `what=text` it will search the text
and tags, with `text=` and `tags=`, if available.

----

The new `funcs.sanitize_name` method is used to clean up emojis
inside a provided string, which is usually a claim name
or a claim title.

This requires the `emoji` package to fully identify the emojis,
otherwise only the emoji flags will be filtered.