print_claims: new module with methods to print claim searches
The `claims.print_claims` method was moved to the new module
`print_claims`, and is now called `print_claims.print_tr_claims`.

This method will print the list of claims obtained
with `claim_search`. It is designed particularly to show
the information of trending claims and generic text searches.

Various columns have fixed sizes in order to align the information
nicely as long as the fields aren't too long.