search_ch_all: methods to search all claims in a channel
The method `ch_search_all_claims` searches the newest 20 pages
that can contain the latest 1000 claims, then moves the block height
to the oldest claim in order to get 20 more pages, and another batch
of 1000 claims.
This cycle is repeated, decreasing the block height, until all claims
have been found in the blockchain data.

The method `ch_search_n_claims` searches only the newest pages
depending on the value of the input `number` parameter.
The `number` determines how many groups of 20 pages, called a `cycle`,
it is going to search, and within each cycle, how many individual `pages`,
up to a maximum of 20.

After the claims have been found, the list of claims is sorted
by release time, filtered to make sure the claims are unique,
and then the list is cut so that the resulting list has `number` claims.
Not every single item has a `release_time` in the metadata,
so for these the value of `timestamp` is used to compute its position
in the list.

The total size of all elements in the resulting list
is also calculated and printed in GiB.

The supporting methods to sort and calculate the size of the claims
are placed in an auxiliary module `search_utils.py`.