search_ch: use the `ch_search_n_claims` method to get claims
In the `search_ch.ch_search_latest` method
the `search_ch_all.ch_search_n_claims` method is used to download
any number of past claims when `number` is 50 or larger.

When number is lower than or equal to 50, it will get the values
by using `ch_search_fifty_claims`.

If `number=0`, it will return all claims
by calling `search_ch_all.ch_search_all_claims`.

In both cases `reverse=True` is passed to the methods
so that the output list contains the newest claims
at the beginning of the list.