print: new parameter `sanitize` for `print_items` to clean up emojis
If `sanitize=True` it will remove the emojis
from the name of the claim, the name of the channel,
and the title of the claim.

This option requires the `emoji` package to be installed.

If the emoji package is not installed, the text won't be cleaned up
but it will not produce an error.

Removing the emojis from the text is necessary in case the output
of `print_items` is used in another application that doesn't
work well showing emojis, such as Tk interfaces.