download: skip the download of collections if `collection=False`
The code to download a collection is placed
in a new method `download_collection`.

This method will only be called if `collection=True`
and if the claim is actually a collection.

If the claim is not a collection, then `download_collection`
will not be called, but neither will the internal `lbrynet_get`.
That is, the download will only proceed if it's an actual
downloadable stream.