print: reorder the variables in `print_items`
To simplify the extraction of information
we use `meta = item["metadata"]`.

We also reorder the printing of the information fields
so that `st_type` (stream type) comes after the number of blobs.
The type is printed as 9 characters long to make sure it aligns
well.

After the type comes `st_path` to display `media` or `no-media`.
This is also displayed in a fixed size. Previously, this field
was displayed at the end, but now it comes in the middle
so it is more prominent but still aligns well.