print: new parameter `blocks` for `print_items`
The `blocks` parameter prints the block height in the blockchain
for the current claim. This gives us an idea of when the claim
was published in the LBRY network. It is related to the `timestamp`.
Higher block heights are newer claims and smaller heights are older claims.

The block height is printed to 8 character so it fits
comfortably up to 99 million blocks.

This is printed before the release time.