print: new parameter `size` for `print_items`
Show the length of the stream in minutes and seconds, like `14:22`,
when possible (audio and video). This may be `00:00` if the claim
is a document, or binary object, etc., that doesn't have a duration.

Also add the size of the stream in mebibytes (MB).
The original size is stored in bytes, but is converted to mebibytes
by dividing by 1024 twice, or 1024^2.