zeed_defaults: improve the printing of channels
When inspecting the `cfg.channels` variable
print only the first three elements if they are found.
If this variable is missing or empty, return from the function,
as this variable is mandatory to start downloading content.

Also inspect the `cfg.never_delete` variable
to print only the first three elements.
This variable is optional so if it does not exist,
or if it exists but is empty, set it to `None`.